Add -Werror to compile GPDB (#1005)
All warning were fixed.
1) checking for not-NULL was added before calling strlen().
2) converting long int to float8 was added for comparison between long int and
float8.
3) Unused variables were removed.
4) Useless memset's were removed from the insertsetbit function, because the buf
variable was zeroed before sending to the insertsetbit function.
5) Assert was fixed in the CreateDXLProjectNullsForGroupingSets function.
Execution of the Find method is extra, because it is checked below in the
condition and it may return nullptr.

Ticket: ADBDEV-5786

+++ src/backend/utils/adt/interpolate.c
@@ -300,7 +300,7 @@ linterp_int64(PG_FUNCTION_ARGS)
 	else 
 	{
 		r = round(y0+p*(y1-y0));
-		if ( r < LONG_MIN || r > LONG_MAX )
+		if ( r < (float8)LONG_MIN || r > (float8)LONG_MAX )
 			ereport(ERROR,
 					(errcode(ERRCODE_NUMERIC_VALUE_OUT_OF_RANGE),
 					 errmsg("value \"%f\" is out of range for type bigint", r)));
